fix: reset f64 to f32

+++ src/camera.rs
@@ -4,7 +4,7 @@ use bevy::render::camera::PerspectiveProjection;
 
 pub struct PanOrbitCamera {
     pub focus: Vec3,
-    pub radius: f64,
+    pub radius: f32,
     pub upside_down: bool,
 }
